requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
protobuf==3.13.0
six==1.15.0
Werkzeug==1.0.1
WTForms==2.3.3
